% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{drop_if_variance_below}
\alias{drop_if_variance_below}
\title{drop variables with variance below a certain threshold}
\usage{
drop_if_variance_below(dataset, threshold)
}
\arguments{
\item{dataset}{A dataframe}

\item{threshold}{A number; variance threshold below which variables}
}
\value{
dataset with same name but with variables with variance below threshold dropped.
}
\description{
drop variables with variance below a certain threshold
}
